requirements.txt
Django==4.2
djangorestframework==3.15.1
python-dotenv==1.0.1
redis==5.0.3
sqlparse==0.4.2
psycopg[binary]==3.1.12
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3
drf-nested-routers==0.93.5
pylint
autopep8
